Move api functionality into a seperate folder
