--- src/api.js
@@ -1,16 +1,0 @@
-const root = 'https://swapi.co/api';
-
-const apiGet = (url) => {
-  return fetch(url).then( response => response.json() );
-}
-
-//export default functions that make api calls
-export default {
-  getMovie(number) {
-    console.log(`here's movie ${number}`)
-  },
-  getData(category) {
-    // return apiGet(`${root}/${category}/`)
-
-  }
-}
